refactor(herosection): shift stuff out
