+++ src/templates/homepage/HeroSection/HeroImageOnlyLayout.tsx
@@ -1,0 +1,3 @@
+export const HeroImageOnlyLayout = () => {
+  return <div className="min-height-mobile" />
+}
